removed redundant arima file
